Design: mydashboard 반응형 UI 구현
